requirements.txt
boto3
faiss-cpu
sentence-transformers
GitPython
langchain
pyyaml
